UCS/TOPO: Use registered Proto providers modules implementing Proto API
