+++ src/ucs/sys/topo/providers/default/ucx-topo-default.pc.in
@@ -1,0 +1,15 @@
+#
+# Copyright (c) 2022, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+#
+# See file LICENSE for terms.
+#
+
+prefix = @prefix@
+exec_prefix = @exec_prefix@
+libdir = @libdir@/ucx
+
+Name: @PACKAGE@-topo-default
+Description: Unified Communication X Library Default Topo Provider module
+Version: @VERSION@
+Libs:
+Libs.private: -L${libdir} -lucs_topo_default -Wl,--undefined=ucs_topo_default_init
